[feat] #57 UiState, Viewmodel 구현

+++ data/src/main/java/com/dpm/sixpack/data/repository/FeedRepositoryImpl.kt
@@ -42,7 +42,7 @@ class FeedRepositoryImpl @Inject constructor(
 
 
     override suspend fun postReaction(
-        selfieId: Int,
+        selfieId: Long,
         emojiType: String
     ): DoRunResult<ReactionResult> =
         withContext(Dispatchers.IO) {
